requirements.txt
python-dotenv==0.21.1
requests==2.22.0
requests-unixsocket==0.2.0